Address more -Wsign-compare issues
Address more signedness/unsignedness issues, like:

../sahara.c: In function ‘sahara_debug64_one’:
../qdl.h:20:12: warning: comparison of integer expressions of different
   signedness: ‘long unsigned int’ and ‘int’ [-Wsign-compare]
   20 |         _x < _y ? _x : _y;      \
      |            ^
../sahara.c:286:26: note: in expansion of macro ‘MIN’
  286 |                 remain = MIN((uint64_t)(region.length - chunk), DEBUG_BLOCK_SIZE);
      |                          ^~~
../qdl.h:20:24: warning: operand of ‘?:’ changes signedness from ‘int’ to
  ‘long unsigned int’ due to unsignedness of other operand [-Wsign-compare]
   20 |         _x < _y ? _x : _y;      \

../gpt.c: In function ‘gpt_find_by_name’:
../gpt.c:255:65: warning: comparison of integer expressions of different
  signedness: ‘unsigned int’ and ‘int’ [-Wsign-compare]
  255 |                 if (*phys_partition >= 0 && gpt_part->partition != *phys_partition)

Now tools are built without any warnings when -Wsign-compare is enabled.

+++ gpt.c
@@ -252,7 +252,7 @@ int gpt_find_by_name(struct qdl_device *qdl, const char *name, int *phys_partiti
 		return -1;
 
 	for (gpt_part = gpt_partitions; gpt_part; gpt_part = gpt_part->next) {
-		if (*phys_partition >= 0 && gpt_part->partition != *phys_partition)
+		if (*phys_partition >= 0 && gpt_part->partition != (unsigned int)(*phys_partition))
 			continue;
 
 		if (strcmp(gpt_part->name, name))
